+++ src/index.js
@@ -3,3 +3,4 @@
 export { exCombineReducers } from './exCombineReducers';
 export { makePlainReducer } from './makePlainReducer';
 export { makeExReducer } from './makeExReducer';
+export { ExReducerDependenciesChanges } from './ExReducerDependenciesChanges';